pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yalehacks"
version = "0.1.0"
description = "Clothing tag vision read, Bright Data search, sustainability LangGraph"
requires-python = ">=3.11"
dependencies = [
  "pymongo>=4.6",
  "python-dotenv>=1.0",
  "Pillow>=10.0",
  "beautifulsoup4>=4.12",
  "langgraph>=0.2",
  "langchain-core>=0.3",
  "httpx>=0.27",
  "groq>=0.11",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "python-multipart>=0.0.9",
  "typer>=0.12",
]

[project.scripts]
yalehacks = "yalehacks.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
